Move border and design modules to top-level.
