+++ pygvisuals/border/compoundborder.py
