+++ pygvisuals/designs/dark.py
