+++ pygvisuals/designs/hologram.py
